+++ provider/pkg/provider/ids.go
@@ -1,4 +1,4 @@
-// Copyright 2016-2021, Pulumi Corporation.
+// Copyright 2016-2023, Pulumi Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -45,22 +45,35 @@ func getDefaultName(
 		return v, nil
 	}
 
+	// Generate naming trivia for the resource.
+	namingTriviaApplies, namingTrivia, err := schema.CheckNamingTrivia(sdkName, news, autoNamingSpec.TriviaSpec)
+	if err != nil {
+		return resource.PropertyValue{}, err
+	}
+
 	// Generate random name that fits the length constraints.
 	name := urn.Name().String()
 	prefix := name + "-"
 	randLength := 7
-	if len(prefix)+randLength < autoNamingSpec.MinLength {
-		randLength = autoNamingSpec.MinLength - len(prefix)
+	if len(prefix)+namingTrivia.Length()+randLength < autoNamingSpec.MinLength {
+		randLength = autoNamingSpec.MinLength - len(prefix) - namingTrivia.Length()
 	}
 
 	maxLength := 0
 	if autoNamingSpec.MaxLength > 0 {
-		left := autoNamingSpec.MaxLength - len(prefix)
+		left := autoNamingSpec.MaxLength - len(prefix) - namingTrivia.Length()
 
 		if left <= 0 {
-			return resource.PropertyValue{}, fmt.Errorf("failed to auto-generate value for %[1]q."+
-				" Prefix: %[2]q is too large to fix max length constraint of %[3]d. Please provide a value for %[1]q",
-				sdkName, prefix, autoNamingSpec.MaxLength)
+			if namingTrivia.Length() > 0 {
+				return resource.PropertyValue{}, fmt.Errorf("failed to auto-generate value for %[1]q."+
+					" Prefix: %[2]q is too large to fix max length constraint of %[3]d"+
+					" with required suffix %[4]q. Please provide a value for %[1]q",
+					sdkName, prefix, autoNamingSpec.MaxLength, namingTrivia.Suffix)
+			} else {
+				return resource.PropertyValue{}, fmt.Errorf("failed to auto-generate value for %[1]q."+
+					" Prefix: %[2]q is too large to fix max length constraint of %[3]d. Please provide a value for %[1]q",
+					sdkName, prefix, autoNamingSpec.MaxLength)
+			}
 		}
 		if left < randLength {
 			randLength = left
@@ -73,5 +86,11 @@ func getDefaultName(
 	if err != nil {
 		return resource.PropertyValue{}, err
 	}
+
+	// Apply naming trivia to the generated name.
+	if namingTriviaApplies {
+		random = namingTrivia.ApplyTrivia(sdkName, random)
+	}
+
 	return resource.NewStringProperty(random), nil
 }